ort-analyzer: Support using a package curation directory
ORT users can correct invalid or missing package metadata and set the
concluded license for packages using a curations. By default ORT will use
a curations.yml file however it also supports using a directory with
multiple curation files is supported, see [1]. As ort-config [2] uses a
curations directory and as it used as a fallback in ORT for GitLab we
need to support both ways of specifying curations.

[1]: https://github.com/oss-review-toolkit/ort/blob/main/docs/config-file-curations-yml.md
[2]: https://github.com/oss-review-toolkit/ort-config

Signed-off-by: Thomas Steenbergen <thomas.steenbergen@here.com>