gitlab-ci: Make Downloader script conditional
Makes executing the Downloader optional to allow use of the default checkout of GitLab. This is usefule when ORT for GitLab is run in a pipeline where the project code is already cloned on pipeline start up.

Note: As GitLab does not allow naming the default checkout directory yet (as option `ci_builds_dir` for runners is not enabled by default, see https://gitlab.com/gitlab-org/gitlab-runner/-/issues/4203), users might to need to change the project directory variable in case the user is not be able to set it to the default 'project'.

Signed-off-by: Josef Andersson <josef.andersson@gmail.com>